% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skew_functions.r
\name{add_skew}
\alias{add_skew}
\title{Skewness}
\usage{
add_skew(fs_samples_df)
}
\arguments{
\item{fs_samples_df}{df of all samples, with columns for sim, year, season, treatment, source}
}
\value{
summary of skewness per sample
}
\description{
Skewness
}
