% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling_functions.R
\name{build_p_table}
\alias{build_p_table}
\title{Build largest necessary P table}
\usage{
build_p_table(list_of_datasets)
}
\arguments{
\item{list_of_datasets}{a list of datasets}
}
\value{
P table that will cover all the datasets
}
\description{
Build largest necessary P table
}
