% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leg_functions.R
\name{get_all_cds}
\alias{get_all_cds}
\title{Get distance to centroid}
\usage{
get_all_cds(fs_samples_df, leg_coeffs_df)
}
\arguments{
\item{fs_samples_df}{df of all samples, with columns for sim, year, season, treatment, source}

\item{leg_coeffs_df}{result of get_all_coeffs}
}
\value{
fs_samples_df with column for distance to centroid
}
\description{
Get distance to centroid
}
