% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leg_functions.R
\name{get_all_coeffs}
\alias{get_all_coeffs}
\title{Legendre coefficients}
\usage{
get_all_coeffs(fs_samples_df, nleg = NULL)
}
\arguments{
\item{fs_samples_df}{df of all samples, with columns for sim, year, season, treatment, source}

\item{nleg}{nb of polynomials to use, defaults to min(10, nspp - 2)}
}
\value{
dataframe of leg coefficients
}
\description{
Legendre coefficients
}
