% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_plant_data.R
\name{get_plant_sad}
\alias{get_plant_sad}
\title{Get Portal plant abundance}
\usage{
get_plant_sad(census_year = 1994, season = "summer",
  plot_treatment = "control")
}
\arguments{
\item{census_year}{census year}

\item{season}{"summer" or "winter"}

\item{plot_treatment}{"control" or "exclosure"}
}
\value{
dataframe of abundance and rank
}
\description{
Get Portal plant abundance
}
